Reorganize directories and file names to make things easier for the user to understand at a glance.
