+++ user/apps/oracle/ai/ensemble.py
@@ -1,3 +1,4 @@
+import __init__
 from typing import Dict, Any, List
 
 import numpy as np
@@ -5,7 +6,6 @@ import pandas as pd
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.preprocessing import RobustScaler
 
-import __init__
 from h1st.model.ensemble.ensemble_modeler import EnsembleModeler
 from h1st.model.model import Model
 from h1st.model.ml_model import MLModel
@@ -69,3 +69,10 @@ class MyEnsembleModeler(EnsembleModeler):
             my_ensemble.stats = self.stats.copy()
             
         return my_ensemble
+
+
+class MyEnsemble(MLModel):
+    def predict(self, data):
+        pass
+
+
