+++ user/apps/oracle/ai/oracle.py
