--- user/model/ensemble.py
@@ -1,10 +1,0 @@
-from typing import Dict, Any
-
-import __init__
-from h1st.model.ml_model import MLModel
-
-class MyEnsemble(MLModel):
-    def predict(self, data):
-        pass
-
-
