--- user/model/oracle.py
@@ -1,11 +1,0 @@
-import __init__
-from h1st.h1flow.h1flow import Graph
-
-
-class MyOracle(Graph):
-    def __init__(self, k_model, k_gen, ensemble):
-        super().__init__()
-        self.start()
-        self.add([k_model, k_gen])
-        self.add(ensemble)
-        self.end()
