+++ user/models/boolean_model.py
